requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.4.0
docopt==0.6.2
idna==3.10
keyboard==0.13.5
MouseInfo==0.1.3
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pefile==2023.2.7
pillow==11.0.0
pipreqs==0.4.13
pushbullet.py==0.12.0
PyAutoGUI==0.9.54
pygame==2.6.1
PyGetWindow==0.0.9
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-magic==0.4.27
pytweening==1.2.0
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==75.2.0
urllib3==2.2.3
websocket-client==1.8.0
yarg==0.1.10
